changes for configurable example to make it wasm compatible
